requirements.txt
Django==3.2.6
djangorestframework
